gesture-dj/agent/requirements.txt
dedalus_labs
python-dotenv
websockets
aiohttp
